Updated to Java 8, removed raw type warnings, and formatted code
